api_yamdb/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-filter==22.1
django-import-export==3.0.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
et-xmlfile==1.1.0
flake8==5.0.4
idna==3.4
importlib-metadata==1.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==22.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tablib==3.3.0
toml==0.10.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
xlrd==2.0.1
xlwt==1.3.0
zipp==3.11.0
python-dotenv==0.21.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
